requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
supabase==2.18.1
bleach==6.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx>=0.26
python-dotenv==1.0.0
email-validator==2.1.0